requirements/binance.txt
# Core API stack
fastapi>=0.118.0,<0.119.0
uvicorn>=0.37.0,<0.38.0
pydantic>=2.12.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Async communication & data access
aiohttp>=3.13.0,<4.0.0
websockets>=12.0
redis[hiredis]>=5.0.0,<7.0.0

# Data processing
pandas>=2.3.0,<2.4.0
numpy>=2.3.0,<2.4.0

# Database layer
SQLAlchemy>=2.0.30,<2.1.0
asyncpg>=0.29.0,<0.31.0
psycopg2-binary>=2.9.0,<3.0.0

# Trading & analytics
backtrader>=1.9.78
yfinance>=0.2.65
matplotlib>=3.8.0,<3.9.0
scipy>=1.12.0,<2.0.0

# Monitoring & Metrics
prometheus-client>=0.19.0,<1.0.0
prometheus-fastapi-instrumentator>=7.0.0,<8.0.0

# Utilities
requests>=2.31.0,<2.32.0
pytz>=2024.1
